pyproject.toml
[project]
name = "pos-ecod-edge-ensemble"
version = "0.1.0"
description = "ECOD real-time detection with IForest complement for POS edge/gateway."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyod>=2.0.5",
  "scikit-learn>=1.5.0",
  "numpy>=2.2.0",
  "scipy>=1.14.0",
  "pandas>=2.2.0",
  "rich>=13.7.0",
  "pyyaml>=6.0.1"
]

[project.scripts]
pos-ecod = "ecod_edge.cli:main"

[tool.uv]
# ensures editable installs are respected
